\name{acde}
\alias{acde}
\title{Extract variance components of twinlm object}
\usage{
  acde(x, ...)
}
\arguments{
  \item{x}{twinlm object}

  \item{...}{Additional arguments parsed on to lower-level
  functions}
}
\description{
  Extract variance components of twinlm object
}
\author{
  Klaus K. Holst
}

